corrected calculation for buffer_delay

+++ src/headers/global.h
@@ -24,7 +24,7 @@ using namespace std;
 
 const int timing_resolution = 100000; // timing accuracy (1000 to 1000000)
 const int micro_multiplier = 1000000 / timing_resolution;
-const int buffer_delay = timing_resolution / (SAMPLE_RATE/BUFF_SIZE); // delay period in microseconds of the audio callabck frequency
+const int buffer_delay = (double)timing_resolution / ((double)SAMPLE_RATE/(double)BUFF_SIZE); // delay period in microseconds of the audio callabck frequency
 
 struct Pallet{ // global color pallet
     SDL_Color black;
